-fixing bugs

+++ loginapp/models.py
@@ -26,7 +26,7 @@ class User(AbstractUser):
     is_superuser = models.SmallIntegerField(default=0)
     is_active = models.SmallIntegerField(default=1)
     deleted = models.SmallIntegerField(default=0)
-    level = models.SmallIntegerField(default=0)
+    level = models.IntegerField(default=0)
 
     @staticmethod
     def get_all_user(user):
